Fix edge case where we can not go get the repo
This fixes an edge case where the repo being tested has a dependency that
has had its git history rebased, thus causing go get to fail, and a subsequent
delete of that repo and another go get to fail as well. Now, if go get fails the first time,
we read the error message and clear the directory that caused the error. Then we try again.
If we get the same error two times in a row, something is wrong, and we return failure. If not,
we keep going and deleting cached repos until no error is returned.