Add option to hide repos from the recently generated list
This option allows users to decide if a repository will be displayed in
the "Recently Generated" box. People might want to use the web service
to check their semi-private or experimental projects but keep their
existence in private. Notice that by default the "Recently Generated"
list serves as a way to spam visitors in order to make them check
non-relevant repositories. I will keep this option as an opt-in instead
of opt-out for backward compatibility.