Fix command for starting local server
It seems that the document is outdated. The `start-dev` command does not seem to exist in the Makefile anymore. Running the command fails:

```
make start-dev
make: *** No rule to make target `start-dev'.  Stop.
```
Is it correct to assume that the `start` command should be used for development and production environment? I change the documentation to reflect that assuming that's the case.